requirements.txt
Flask
chromadb
openai
python-dotenv
